[minishell] ADD: env 새로 작업 시작 (이전것은 백업)
